Updated mechanism for the generation of random numbers
- Changed the type of random numbers from u32 to u8
- Made default generator `os`
- Changed the method of generating random numbers on fill_bytes
- Added banch for generators

+++ examples/complex.rs
@@ -3,12 +3,12 @@ extern crate nanoid;
 
 use rand::Rng;
 
-fn random (size: usize) -> Vec<u32> {
+fn random (size: usize) -> Vec<u8> {
     let mut rng = rand::thread_rng();
-    let mut bytes: Vec<u32> = vec![0; size];
+    let mut bytes: Vec<u8> = vec![0; size];
 
     for i in 0..size {
-        bytes[i] = rng.gen::<u32>();
+        bytes[i] = rng.gen::<u8>();
     }
 
     bytes
